fix: update cluster for config reloading using watcher
