+++ cluster/service/cluster_blackbox_test.go
@@ -11,6 +11,12 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"testing"
+	"os"
+	"io/ioutil"
+	"io"
+	"github.com/fabric8-services/fabric8-cluster/configuration"
+	"time"
+	"github.com/fabric8-services/fabric8-cluster/gormapplication"
 )
 
 func TestCluster(t *testing.T) {
@@ -36,6 +42,97 @@ func (s *ClusterServiceTestSuite) TestCreateOrSaveOSOClusterOK() {
 	assert.Len(s.T(), clusters, 3)
 
 	verifyClusters(s.T(), clusters)
+}
+
+func (s *ClusterServiceTestSuite) TestClusterConfigurationWatcher() {
+	t := s.T()
+	// Create a temp file with content from ./conf-files/oso-clusters-custom.conf
+	tmpFileName := createTempClusterConfigFile(t)
+	defer os.Remove(tmpFileName)
+
+	// Load configuration from the temp file
+	config, err := configuration.NewConfigurationData("", tmpFileName)
+	require.NoError(t, err)
+	cluster := config.GetOSOClusterByURL("https://api.starter-us-east-2a.openshift.com")
+	require.NotNil(t, cluster)
+
+	original := cluster.CapacityExhausted
+
+	// initialize application with new config
+	application := gormapplication.NewGormDB(s.DB, config)
+	// Start watching
+	haltWatcher, err := application.ClusterService().InitializeClusterWatcher()
+	require.NoError(t, err)
+	defer haltWatcher()
+
+	// Update the config file
+	updateClusterConfigFile(t, tmpFileName, "./configuration/conf-files/tests/oso-clusters-capacity-updated.conf")
+	// Check if it has been updated
+	waitForConfigUpdate(t, config, !original)
+
+	// Update the config file back to the original
+	updateClusterConfigFile(t, tmpFileName, "./configuration/conf-files/oso-clusters.conf")
+	// Check if it has been updated
+	waitForConfigUpdate(t, config, original)
+
+	// Update the config file to some invalid data
+	updateClusterConfigFile(t, tmpFileName, "./configuration/conf-files/tests/oso-clusters-invalid.conf")
+	// The configuration should not change
+	waitForConfigUpdate(t, config, original)
+	updateClusterConfigFile(t, tmpFileName, "./configuration/conf-files/tests/oso-clusters-empty.conf")
+	// The configuration should not change
+	waitForConfigUpdate(t, config, original)
+
+	// Fix the file
+	updateClusterConfigFile(t, tmpFileName, "./configuration/conf-files/tests/oso-clusters-capacity-updated.conf")
+	// Now configuration should be updated
+	waitForConfigUpdate(t, config, !original)
+}
+
+func (s *ClusterServiceTestSuite) TestClusterConfigurationWatcherNoErrorForDefaultConfig() {
+	s.Application = gormapplication.NewGormDB(s.DB, s.Configuration)
+	haltWatcher, err := s.Application.ClusterService().InitializeClusterWatcher()
+	require.NoError(s.T(), err)
+	defer haltWatcher()
+}
+
+func createTempClusterConfigFile(t *testing.T) string {
+	to, err := ioutil.TempFile("", "oso-clusters.conf")
+	require.NoError(t, err)
+	defer to.Close()
+
+	from, err := os.Open("./../../configuration/conf-files/oso-clusters.conf")
+	require.NoError(t, err)
+	defer from.Close()
+
+	_, err = io.Copy(to, from)
+	require.NoError(t, err)
+	return to.Name()
+}
+
+func updateClusterConfigFile(t *testing.T, to, from string) {
+	fromFile, err := os.Open("./../../" + from)
+	require.NoError(t, err)
+	defer fromFile.Close()
+
+	toFile, err := os.OpenFile(to, os.O_RDWR|os.O_CREATE, 0666)
+	require.NoError(t, err)
+	defer toFile.Close()
+
+	_, err = io.Copy(toFile, fromFile)
+	require.NoError(t, err)
+}
+
+func waitForConfigUpdate(t *testing.T, config *configuration.ConfigurationData, expected bool) {
+	for i := 0; i < 30; i++ {
+		time.Sleep(100 * time.Millisecond)
+		cluster := config.GetOSOClusterByURL("https://api.starter-us-east-2a.openshift.com")
+		require.NotNil(t, cluster)
+		if expected == cluster.CapacityExhausted {
+			return
+		}
+	}
+	require.Fail(t, "cluster config has not been reloaded within 3s")
 }
 
 func verifyClusters(t *testing.T, clusters []repository.Cluster) {
