+++ configuration/configuration.go
@@ -13,7 +13,6 @@ import (
 	commoncfg "github.com/fabric8-services/fabric8-common/configuration"
 	"github.com/fabric8-services/fabric8-common/httpsupport"
 
-	"github.com/fsnotify/fsnotify"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
@@ -274,7 +273,7 @@ func readFromJSONFile(configFilePath string, defaultConfigFilePath string, confi
 	} else {
 		// If the JSON configuration file has not been specified
 		// then we default to <defaultConfigFile>
-		configFilePath, err = pathExists(defaultConfigFilePath)
+		configFilePath, err = PathExists(defaultConfigFilePath)
 		if err != nil {
 			return nil, nil, defaultConfigFilePath, err
 		}
@@ -315,7 +314,7 @@ func (c *ConfigurationData) appendDefaultConfigErrorMessage(message string) {
 	}
 }
 
-func pathExists(pathToCheck string) (string, error) {
+func PathExists(pathToCheck string) (string, error) {
 	_, err := os.Stat(pathToCheck)
 	if err == nil {
 		return pathToCheck, nil
@@ -336,76 +335,8 @@ func getOSOClusterConfigFile() string {
 	return envOSOClusterConfigFile
 }
 
-// InitializeClusterWatcher initializes a file watcher for the cluster config file
-// When the file is updated the configuration synchronously reload the cluster configuration
-func (c *ConfigurationData) InitializeClusterWatcher() (func() error, error) {
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		return nil, err
-	}
 
-	go func() {
-		for {
-			select {
-			case event, ok := <-watcher.Events:
-				if !ok {
-					return
-				}
-				if event.Op&fsnotify.Remove == fsnotify.Remove {
-					time.Sleep(1 * time.Second) // Wait for one second before re-adding and reloading. It might be needed if the file is removed and then re-added in some environments
-					err = watcher.Add(event.Name)
-					if err != nil {
-						log.WithFields(map[string]interface{}{
-							"file": event.Name,
-						}).Errorln("cluster config was removed but unable to re-add it to watcher")
-					}
-				}
-				if event.Op&fsnotify.Write == fsnotify.Write || event.Op&fsnotify.Remove == fsnotify.Remove {
-					// Reload config if operation is Write or Remove.
-					// Both can be part of file update depending on environment and actual operation.
-					err := c.reloadClusterConfig()
-					if err != nil {
-						// Do not crash. Log the error and keep using the existing configuration
-						log.WithFields(map[string]interface{}{
-							"err":  err,
-							"file": event.Name,
-							"op":   event.Op.String(),
-						}).Errorln("unable to reload cluster config file")
-					} else {
-						log.WithFields(map[string]interface{}{
-							"file": event.Name,
-							"op":   event.Op.String(),
-						}).Infoln("cluster config file modified and reloaded")
-					}
-				}
-			case err, ok := <-watcher.Errors:
-				if !ok {
-					return
-				}
-				log.WithFields(map[string]interface{}{
-					"err": err,
-				}).Errorln("cluster config file watcher error")
-			}
-		}
-	}()
-
-	configFilePath, err := pathExists(c.clusterConfigFilePath)
-	if err == nil && configFilePath != "" {
-		err = watcher.Add(configFilePath)
-		log.WithFields(map[string]interface{}{
-			"file": c.clusterConfigFilePath,
-		}).Infoln("cluster config file watcher initialized")
-	} else {
-		// OK in Dev Mode
-		log.WithFields(map[string]interface{}{
-			"file": c.clusterConfigFilePath,
-		}).Warnln("cluster config file watcher not initialized for non-existent file")
-	}
-
-	return watcher.Close, err
-}
-
-func (c *ConfigurationData) reloadClusterConfig() error {
+func (c *ConfigurationData) ReloadClusterConfig() error {
 	c.mux.Lock()
 	defer c.mux.Unlock()
 
@@ -483,6 +414,11 @@ func (c *ConfigurationData) GetOSOClusterByURL(url string) *OSOCluster {
 	return nil
 }
 
+// GetOSOConfigurationFilePath returns the oso configuration file path.
+func (c *ConfigurationData) GetOSOConfigurationFilePath() string {
+	return c.clusterConfigFilePath
+}
+
 // GetDefaultConfigurationFile returns the default configuration file.
 func (c *ConfigurationData) GetDefaultConfigurationFile() string {
 	return defaultConfigFile
@@ -533,7 +469,7 @@ func (c *ConfigurationData) setConfigDefaults() {
 
 	c.v.SetDefault(varLogLevel, defaultLogLevel)
 
-	// By default, test data should be cleaned from DB, unless explicitely said otherwise.
+	// By default, test data should be cleaned from DB, unless explicitly said otherwise.
 	c.v.SetDefault(varCleanTestDataEnabled, true)
 	// By default, DB logs are not output in the console
 	c.v.SetDefault(varDBLogsEnabled, false)
