+++ main.go
@@ -18,7 +18,6 @@ import (
 	"github.com/fabric8-services/fabric8-common/log"
 	"github.com/fabric8-services/fabric8-common/token"
 
-	"context"
 	"github.com/fabric8-services/fabric8-cluster/gormapplication"
 	"github.com/goadesign/goa"
 	"github.com/goadesign/goa/logging/logrus"
@@ -28,6 +27,7 @@ import (
 	"github.com/jinzhu/gorm"
 	_ "github.com/lib/pq"
 	"github.com/prometheus/client_golang/prometheus"
+	"context"
 )
 
 func main() {
@@ -95,14 +95,6 @@ func main() {
 	}
 	defer haltSentry()
 
-	// Initialize cluster config watcher
-	haltWatcher, err := config.InitializeClusterWatcher()
-	if err != nil {
-		log.Panic(nil, map[string]interface{}{
-			"err": err,
-		}, "failed to setup the cluster config watcher")
-	}
-	defer haltWatcher()
 
 	if config.DeveloperModeEnabled() && log.IsDebug() {
 		db = db.Debug()
@@ -149,6 +141,21 @@ func main() {
 	// Create DB
 	appDB := gormapplication.NewGormDB(db, config)
 
+	// Create OSO cluster config for the first time
+	if err := appDB.ClusterService().CreateOrSaveOSOClusterFromConfig(context.Background()); err != nil {
+		log.Panic(context.TODO(), map[string]interface{}{
+			"err": err,
+		}, "failed to create or save cluster")
+	}
+	// Initialize cluster config watcher
+	haltWatcher, err := appDB.ClusterService().InitializeClusterWatcher()
+	if err != nil {
+		log.Panic(context.TODO(), map[string]interface{}{
+			"err": err,
+		}, "failed to setup the cluster config watcher")
+	}
+	defer haltWatcher()
+
 	// Setup Security
 	tokenManager, err := token.DefaultManager(config)
 	if err != nil {
@@ -169,7 +176,6 @@ func main() {
 	app.MountStatusController(service, statusCtrl)
 
 	// Mount "clusters" controller
-	appDB.ClusterService().CreateOrSaveOSOClusterFromConfig(context.Background())
 	clustersCtrl := controller.NewClustersController(service, config)
 	app.MountClustersController(service, clustersCtrl)
